correct signature for dbl_alltoallv_sparse in mpispike.c

+++ src/nrnmpi/mpispike.c
@@ -309,10 +309,9 @@ static int MPI_Alltoallv_sparse(void *sendbuf, int *sendcnts, int *sdispls,
 }
 
 
-extern void nrnmpi_dbl_alltoallv_sparse(int* s, int* scnt, int* sdispl, int* r, int* rcnt, int* rdispl) {
+extern void nrnmpi_dbl_alltoallv_sparse(double* s, int* scnt, int* sdispl,  double* r, int* rcnt, int* rdispl) {
         MPI_Alltoallv_sparse(s, scnt, sdispl, MPI_DOUBLE, r, rcnt, rdispl, MPI_DOUBLE, nrnmpi_comm);
 }
-
 extern void nrnmpi_int_alltoallv_sparse(int* s, int* scnt, int* sdispl, int* r, int* rcnt, int* rdispl) {
 	MPI_Alltoallv_sparse(s, scnt, sdispl, MPI_INT, r, rcnt, rdispl, MPI_INT, nrnmpi_comm);
 }
